refactor: create shared-helpers package
