+++ package.json
@@ -7,7 +7,7 @@
     "test": "yarn test:v4 && yarn test:v5 && yarn test:v6",
     "test:v4": "mocha packages/eslint-plugin-pf-codemods/test/rules/v4",
     "test:v5": "mocha packages/eslint-plugin-pf-codemods/test/rules/v5",
-    "test:v6": "yarn build && mocha packages/eslint-plugin-pf-codemods/dist/js/packages/eslint-plugin-pf-codemods/src/rules/v6/*/*.test.js",
+    "test:v6": "yarn build && mocha packages/eslint-plugin-pf-codemods/dist/js/rules/v6/*/*.test.js",
     "test:v4:single": "pf-codemods --v4 --no-cache test/v4test.tsx",
     "test:v5:single": "pf-codemods --no-cache test/test.tsx",
     "test:v6:single": "yarn build && pf-codemods --v6 --no-cache packages/eslint-plugin-pf-codemods/src/rules/v6/*/*Input.tsx",
