+++ packages/css-vars-updater/package.json
@@ -4,9 +4,7 @@
   "description": "Utility to update Patternfly CSS variables when migrating from V5 to a new design token system in V6",
   "author": "Red Hat",
   "license": "MIT",
-  "bin": {
-    "css-vars-updater": "./dist/packages/css-vars-updater/src/cli.js"
-  },
+  "bin": "./dist/cli.js",
   "engines": {
     "node": ">=18"
   },
@@ -22,7 +20,8 @@
     "commander": "^5.1.0",
     "diff": "^5.1.0",
     "glob": "^10.2.6",
-    "inquirer": "^12.0.1"
+    "inquirer": "^12.0.1",
+    "shared-helpers": "^1.0.0"
   },
   "devDependencies": {
     "@types/diff": "^5.0.9",
