+++ packages/eslint-plugin-pf-codemods/src/rules/v6/tokensPrefixWithT/tokens-prefix-with-t.ts
@@ -8,7 +8,7 @@ import {
   getImportPath,
 } from "../../helpers";
 import { Identifier, ImportSpecifier } from "estree-jsx";
-import { tokensToPrefixWithT } from "../../../../../../helpers/tokenLists";
+import { tokensToPrefixWithT } from "shared-helpers/dist/tokenLists";
 
 module.exports = {
   meta: { fixable: "code" },
