+++ packages/shared-helpers/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "shared-helpers",
+  "version": "1.0.0",
+  "description": "Shared helpers for the monorepo",
+  "scripts": {
+    "build": "tsc --build --verbose ./tsconfig.json",
+    "clean": "rimraf ./dist"
+  },
+  "author": "Red Hat",
+  "license": "MIT",
+  "dependencies": {
+    "colors": "^1.4.0",
+    "diff": "^5.1.0"
+  },
+  "devDependencies": {
+    "@types/diff": "^5.0.9",
+    "typescript": "^5.4.2"
+  }
+}
