+++ packages/shared-helpers/src/tokenLists/tokensToPrefixWithT.ts
